make put and delete requests using protected private routes

+++ my-contacts-backend/controllers/contactController.js
@@ -54,6 +54,11 @@ const updateContact = asyncHandler(async (req, res) => {
     throw new Error("Contact not found");
   }
 
+  if (contact.user_id.toString() !== req.user_id) {
+    res.status(403);
+    throw new Error("User dont have permission to update other user contacts");
+  }
+
   const updatedContact = await Contact.findByIdAndUpdate(
     req.params.id,
     req.body,
@@ -66,15 +71,33 @@ const updateContact = asyncHandler(async (req, res) => {
 //@route DELETE api/contacts/:id
 //@access private
 
+// const deleteContact = asyncHandler(async (req, res) => {
+//   const contact = await Contact.findById(req.params.id);
+//   if (!contact) {
+//     res.status(404);
+//     throw new Error("Contact not found");
+//   }
+//   if (contact.user_id.toString() !== req.user_id) {
+//     res.status(403);
+//     throw new Error("User dont have permission to update other user contacts");
+//   }
+
+//   await Contact.deleteOne({ _id: req.params.id }); // Remove the specific contact
+
+//   res.status(200).json(contact);
+// });
+
 const deleteContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
   if (!contact) {
     res.status(404);
     throw new Error("Contact not found");
   }
-
-  await Contact.deleteOne({ _id: req.params.id }); // Remove the specific contact
-
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error("User don't have permission to update other user contacts");
+  }
+  await Contact.deleteOne({ _id: req.params.id });
   res.status(200).json(contact);
 });
